Poly_commitment: move CamlOpeningProof into eval_proof where defined
It is more logical to move where OpeningProof is defined.
This way, it follows the convention.
