+++ poly-commitment/src/commitment.rs
@@ -966,9 +966,9 @@ pub fn inner_prod<F: Field>(xs: &[F], ys: &[F]) -> F {
 
 #[cfg(feature = "ocaml_types")]
 pub mod caml {
-    use super::*;
-
     // polynomial commitment
+    use super::PolyComm;
+    use ark_ec::AffineRepr;
 
     #[derive(Clone, Debug, ocaml::IntoValue, ocaml::FromValue, ocaml_gen::Struct)]
     pub struct CamlPolyComm<CamlG> {
@@ -1035,60 +1035,6 @@ pub mod caml {
             PolyComm {
                 //FIXME something with as_ref()
                 elems: camlpolycomm.unshifted.iter().map(Into::into).collect(),
-            }
-        }
-    }
-
-    // opening proof
-
-    #[derive(ocaml::IntoValue, ocaml::FromValue, ocaml_gen::Struct)]
-    pub struct CamlOpeningProof<G, F> {
-        /// vector of rounds of L & R commitments
-        pub lr: Vec<(G, G)>,
-        pub delta: G,
-        pub z1: F,
-        pub z2: F,
-        pub sg: G,
-    }
-
-    impl<G, CamlF, CamlG> From<OpeningProof<G>> for CamlOpeningProof<CamlG, CamlF>
-    where
-        G: AffineRepr,
-        CamlG: From<G>,
-        CamlF: From<G::ScalarField>,
-    {
-        fn from(opening_proof: OpeningProof<G>) -> Self {
-            Self {
-                lr: opening_proof
-                    .lr
-                    .into_iter()
-                    .map(|(g1, g2)| (CamlG::from(g1), CamlG::from(g2)))
-                    .collect(),
-                delta: CamlG::from(opening_proof.delta),
-                z1: opening_proof.z1.into(),
-                z2: opening_proof.z2.into(),
-                sg: CamlG::from(opening_proof.sg),
-            }
-        }
-    }
-
-    impl<G, CamlF, CamlG> From<CamlOpeningProof<CamlG, CamlF>> for OpeningProof<G>
-    where
-        G: AffineRepr,
-        CamlG: Into<G>,
-        CamlF: Into<G::ScalarField>,
-    {
-        fn from(caml: CamlOpeningProof<CamlG, CamlF>) -> Self {
-            Self {
-                lr: caml
-                    .lr
-                    .into_iter()
-                    .map(|(g1, g2)| (g1.into(), g2.into()))
-                    .collect(),
-                delta: caml.delta.into(),
-                z1: caml.z1.into(),
-                z2: caml.z2.into(),
-                sg: caml.sg.into(),
             }
         }
     }
